working scipy.differential_evolution fitting w/ dask
required various wrappers and special calls to load data everywhere.

specifically, the dast_client.submit(load_model), and
dask_client.replicate(..) calls are crucial for loading, and persisting,
the full model class across all workers. This makes it so the Phenograss()
class, along with all data inside it, isn't copied with every iteration.

dask_scipy_wrapper connects the internal scipy map() with dask, and
chunks up different parameter sets so that workers have more than 1
to work on. This is beneficial as the model is evaluated in ~0.1s.